+++ src/main/java/com/naocraftlab/foggypalegarden/domain/model/FogCharacteristics.java
@@ -1,12 +1,14 @@
 package com.naocraftlab.foggypalegarden.domain.model;
 
 import lombok.Builder;
+import net.minecraft.client.render.FogShape;
 
 @Builder
 public record FogCharacteristics(
         float startDistance,
         float endDistance,
         Color color,
+        FogShape shape,
         float fogDensity
 ) {
 }