+++ src/main/java/com/naocraftlab/foggypalegarden/mixin/PaleGardenFogMixin.java
@@ -1,16 +1,19 @@
 package com.naocraftlab.foggypalegarden.mixin;
 
+import com.naocraftlab.foggypalegarden.config.ConfigManager;
 import com.naocraftlab.foggypalegarden.domain.model.Color;
 import com.naocraftlab.foggypalegarden.domain.model.Environment;
 import com.naocraftlab.foggypalegarden.domain.model.FogCharacteristics;
 import com.naocraftlab.foggypalegarden.domain.model.Weather;
 import com.naocraftlab.foggypalegarden.domain.service.FogService;
 import lombok.val;
+import net.minecraft.client.network.ClientPlayerEntity;
 import net.minecraft.client.render.BackgroundRenderer;
 import net.minecraft.client.render.BackgroundRenderer.FogType;
 import net.minecraft.client.render.Camera;
 import net.minecraft.client.render.Fog;
 import net.minecraft.client.world.ClientWorld;
+import net.minecraft.world.GameMode;
 import net.minecraft.world.RaycastContext;
 import org.joml.Vector4f;
 import org.spongepowered.asm.mixin.Mixin;
@@ -22,7 +25,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import static com.naocraftlab.foggypalegarden.domain.model.Weather.CLEAR;
 import static com.naocraftlab.foggypalegarden.domain.model.Weather.RAIN;
 import static com.naocraftlab.foggypalegarden.domain.model.Weather.THUNDER;
-import static net.minecraft.client.render.FogShape.SPHERE;
 import static net.minecraft.world.LightType.SKY;
 import static net.minecraft.world.RaycastContext.FluidHandling.NONE;
 import static net.minecraft.world.RaycastContext.ShapeType.COLLIDER;
@@ -43,7 +45,12 @@ public abstract class PaleGardenFogMixin {
             float tickDelta,
             CallbackInfoReturnable<Fog> cir
     ) {
-        val focusedEntity = camera.getFocusedEntity();
+        val focusedEntity = (ClientPlayerEntity) camera.getFocusedEntity();
+        val gameMode = resolveGameMode(focusedEntity);
+        if (ConfigManager.currentConfig().getNoFogGameModes().contains(gameMode)) {
+            fogDensity = 0.0f;
+            return;
+        }
         val world = (ClientWorld) focusedEntity.getWorld();
         val blockPos = camera.getBlockPos();
         val biomeEntry = world.getBiome(blockPos);
@@ -53,7 +60,9 @@ public abstract class PaleGardenFogMixin {
 
         val fogCharacteristics = FogService.calculateFogCharacteristics(
                 Environment.builder()
+                        .dimension(world.getRegistryKey().getValue().toString())
                         .biome(biomeEntry.getIdAsString())
+                        .biomeTemperature(biomeEntry.value().getTemperature())
                         .difficulty(world.getDifficulty())
                         .weather(resolveWeather(world))
                         .timeOfDay(world.getTimeOfDay())
@@ -87,6 +96,19 @@ public abstract class PaleGardenFogMixin {
     }
 
     @Unique
+    private static GameMode resolveGameMode(ClientPlayerEntity player) {
+        val attributes = player.getAbilities();
+        if (attributes.creativeMode) {
+            return GameMode.CREATIVE;
+        } else if (attributes.allowFlying && attributes.invulnerable && attributes.flying) {
+            return GameMode.SPECTATOR;
+        } else if (attributes.allowModifyWorld) {
+            return GameMode.SURVIVAL;
+        }
+        return GameMode.ADVENTURE;
+    }
+
+    @Unique
     private static Weather resolveWeather(ClientWorld world) {
         if (world.isThundering()) {
             return THUNDER;
@@ -107,7 +129,7 @@ public abstract class PaleGardenFogMixin {
         return new Fog(
                 fogCharacteristics.startDistance(),
                 fogCharacteristics.endDistance(),
-                SPHERE,
+                fogCharacteristics.shape(),
                 fogCharacteristics.color().red(),
                 fogCharacteristics.color().green(),
                 fogCharacteristics.color().blue(),