+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Condition for applying fog based on the dimension the player is in
 - Condition for applying fog based on the temperature of the biome the player is in
+- Fog shape control (available options: SPHERE and CYLINDER)
 
 ## 1.21.2-2.0.0